#10 リファクタリング
