+++ package.json
@@ -24,6 +24,7 @@
     "@testing-library/dom": "^10.4.0",
     "@types/dompurify": "^3.0.5",
     "dompurify": "^3.1.6",
+    "firebase": "^10.13.0",
     "marked": "^14.0.0",
     "next": "14.2.5",
     "openai": "^4.54.0",
