+++ src/features/register/components/page/RegisterPage.tsx
@@ -1,0 +1,13 @@
+import AuthLayout from '@/components/AuthLayout';
+import { RegisterForm } from '@/features/register/components/ui/RegisterForm';
+import React from 'react';
+
+const RegisterPage = () => {
+	return (
+		<AuthLayout>
+			<RegisterForm />
+		</AuthLayout>
+	);
+};
+
+export default RegisterPage;
